requirements.txt
ollama
beautifulsoup4
langchain
langchain-community
langchain-core
chromadb
google-generativeai
gradio